// src/generated/pistonpanel/instance.ts
export interface InstanceUser {
  /**
   * @generated from protobuf field: string id = 1;
   */
  id: string;
  /**
   * @generated from protobuf field: string username = 2;
   */
  username: string;
  /**
   * @generated from protobuf field: string email = 3;
   */
  email: string;
}
// ...
export interface InstanceListResponse {
  /**
   * @generated from protobuf field: repeated pistonpanel.v1.InstanceListResponse.Instance instances = 1;
   */
  instances: InstanceListResponse_Instance[];
}
/**
 * @generated from protobuf message pistonpanel.v1.InstanceListResponse.Instance
 */
export interface InstanceListResponse_Instance {
  /**
   * @generated from protobuf field: string id = 1;
   */
  id: string;
  /**
   * @generated from protobuf field: string friendly_name = 2;
   */
  friendlyName: string;
  /**
   * @generated from protobuf field: string icon = 5;
   */
  icon: string;
  /**
   * @generated from protobuf field: pistonpanel.v1.InstanceState state = 3;
   */
  state: InstanceState;
  /**
   * @generated from protobuf field: repeated pistonpanel.v1.InstancePermissionState instance_permissions = 4;
   */
  instancePermissions: InstancePermissionState[];
}
/**
 * @generated from protobuf message pistonpanel.v1.InstancePermissionState
 */
export interface InstancePermissionState {
  /**
   * @generated from protobuf field: pistonpanel.v1.InstancePermission instance_permission = 1;
   */
  instancePermission: InstancePermission;
  /**
   * @generated from protobuf field: bool granted = 2;
   */
  granted: boolean;
}
// ...
export interface InstanceAuditLogResponse {
  /**
   * @generated from protobuf field: repeated pistonpanel.v1.InstanceAuditLogResponse.AuditLogEntry entry = 1;
   */
  entry: InstanceAuditLogResponse_AuditLogEntry[];
}
/**
 * @generated from protobuf message pistonpanel.v1.InstanceAuditLogResponse.AuditLogEntry
 */
export interface InstanceAuditLogResponse_AuditLogEntry {
  /**
   * @generated from protobuf field: string id = 1;
   */
  id: string;
  /**
   * @generated from protobuf field: pistonpanel.v1.InstanceUser user = 2;
   */
  user?: InstanceUser;
  /**
   * @generated from protobuf field: pistonpanel.v1.InstanceAuditLogResponse.AuditLogEntryType type = 3;
   */
  type: InstanceAuditLogResponse_AuditLogEntryType;
  /**
   * @generated from protobuf field: google.protobuf.Timestamp timestamp = 4;
   */
  timestamp?: Timestamp;
  /**
   * @generated from protobuf field: string data = 5;
   */
  data: string;
}
/**
 * @generated from protobuf enum pistonpanel.v1.InstanceAuditLogResponse.AuditLogEntryType
 */
export enum InstanceAuditLogResponse_AuditLogEntryType {
  /**
   * @generated from protobuf enum value: EXECUTE_COMMAND = 0;
   */
  EXECUTE_COMMAND = 0,
  /**
   * @generated from protobuf enum value: START_ATTACK = 1;
   */
  START_ATTACK = 1,
  /**
   * @generated from protobuf enum value: PAUSE_ATTACK = 2;
   */
  PAUSE_ATTACK = 2,
  /**
   * @generated from protobuf enum value: RESUME_ATTACK = 3;
   */
  RESUME_ATTACK = 3,
  /**
   * @generated from protobuf enum value: STOP_ATTACK = 4;
   */
  STOP_ATTACK = 4,
}
/**
 * @generated from protobuf enum pistonpanel.v1.InstanceState
 */
export enum InstanceState {
  /**
   * @generated from protobuf enum value: STARTING = 0;
   */
  STARTING = 0,
  /**
   * @generated from protobuf enum value: RUNNING = 1;
   */
  RUNNING = 1,
  /**
   * @generated from protobuf enum value: PAUSED = 2;
   */
  PAUSED = 2,
  /**
   * @generated from protobuf enum value: STOPPING = 3;
   */
  STOPPING = 3,
  /**
   * @generated from protobuf enum value: STOPPED = 4;
   */
  STOPPED = 4,
}
// ...
export const InstanceUser = new InstanceUser$Type();
// ...
class InstanceListResponse$Type extends MessageType<InstanceListResponse> {
  constructor() {
    super('pistonpanel.v1.InstanceListResponse', [
      {
        no: 1,
        name: 'instances',
        kind: 'message',
        repeat: 1 /*RepeatType.PACKED*/,
        T: () => InstanceListResponse_Instance,
      },
    ]);
  }
}
/**
 * @generated MessageType for protobuf message pistonpanel.v1.InstanceListResponse
 */
export const InstanceListResponse = new InstanceListResponse$Type();
// @generated message type with reflection information, may provide speed optimized methods
class InstanceListResponse_Instance$Type extends MessageType<InstanceListResponse_Instance> {
  constructor() {
    super('pistonpanel.v1.InstanceListResponse.Instance', [
      { no: 1, name: 'id', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
      {
        no: 2,
        name: 'friendly_name',
        kind: 'scalar',
        T: 9 /*ScalarType.STRING*/,
      },
      { no: 5, name: 'icon', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
      {
        no: 3,
        name: 'state',
        kind: 'enum',
        T: () => ['pistonpanel.v1.InstanceState', InstanceState],
      },
      {
        no: 4,
        name: 'instance_permissions',
        kind: 'message',
        repeat: 1 /*RepeatType.PACKED*/,
        T: () => InstancePermissionState,
      },
    ]);
  }
}
/**
 * @generated MessageType for protobuf message pistonpanel.v1.InstanceListResponse.Instance
 */
export const InstanceListResponse_Instance =
  new InstanceListResponse_Instance$Type();
// ...
export const InstancePermissionState = new InstancePermissionState$Type();
// ...
class InstanceAuditLogResponse$Type extends MessageType<InstanceAuditLogResponse> {
  constructor() {
    super('pistonpanel.v1.InstanceAuditLogResponse', [
      {
        no: 1,
        name: 'entry',
        kind: 'message',
        repeat: 1 /*RepeatType.PACKED*/,
        T: () => InstanceAuditLogResponse_AuditLogEntry,
      },
    ]);
  }
}
/**
 * @generated MessageType for protobuf message pistonpanel.v1.InstanceAuditLogResponse
 */
export const InstanceAuditLogResponse = new InstanceAuditLogResponse$Type();
// @generated message type with reflection information, may provide speed optimized methods
class InstanceAuditLogResponse_AuditLogEntry$Type extends MessageType<InstanceAuditLogResponse_AuditLogEntry> {
  constructor() {
    super('pistonpanel.v1.InstanceAuditLogResponse.AuditLogEntry', [
      { no: 1, name: 'id', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
      { no: 2, name: 'user', kind: 'message', T: () => InstanceUser },
      {
        no: 3,
        name: 'type',
        kind: 'enum',
        T: () => [
          'pistonpanel.v1.InstanceAuditLogResponse.AuditLogEntryType',
          InstanceAuditLogResponse_AuditLogEntryType,
        ],
      },
      { no: 4, name: 'timestamp', kind: 'message', T: () => Timestamp },
      { no: 5, name: 'data', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
    ]);
  }
}
/**
 * @generated MessageType for protobuf message pistonpanel.v1.InstanceAuditLogResponse.AuditLogEntry
 */
export const InstanceAuditLogResponse_AuditLogEntry =
  new InstanceAuditLogResponse_AuditLogEntry$Type();